+++ IMiPhone/IMNWSocketConnect.m
@@ -104,7 +104,7 @@ char cryptKey[17];
         [data getBytes:bytesToDecode range:NSMakeRange(2, data.length - 3)];
         cryptKey[0] = dataBytes[0];
         cryptKey[1] = dataBytes[1];
-        bitDecode(bytesToDecode , data.length - 3, cryptKey, 16);
+        bitDecode(bytesToDecode , (int)data.length - 3, cryptKey, 16);
         dataDecoded = [NSData dataWithBytes:bytesToDecode length:data.length - 3];
     }
     else {
@@ -136,12 +136,12 @@ char cryptKey[17];
     }
     else if (self.socket.isConnected) {
         NSString *content = [[NSString alloc] initWithData:data encoding:NSUTF8StringEncoding];
-        NSLog(@"Socket Send: %@ , length: %i", content, data.length);
+        NSLog(@"Socket Send: %@ , length: %lu", content, (unsigned long)data.length);
         if (CRYPT) {
             cryptKey[0] = rand() % 127 + 1;
             cryptKey[1] = rand() % 127 + 1;
             char *dataBytes = (char*)[data bytes];
-            bitEncode(dataBytes, data.length, cryptKey, 16);
+            bitEncode(dataBytes, (int)data.length, cryptKey, 16);
             NSMutableData *dataEncoded = [NSMutableData data];
             [dataEncoded appendBytes:cryptKey length:2];
             [dataEncoded appendBytes:dataBytes length:data.length];
